Move date and time formatting to utils package
